refactor: se creo un hook para controlar el pomodoro
